+++ packages/expo-localization/src/ExpoLocalization.ts
@@ -17,7 +17,7 @@ const getNavigatorLocales = () => {
 type ExtendedLocale = Intl.Locale &
   // typescript definitions for navigator language don't include some modern Intl properties
   Partial<{
-    textInfo: { direction: 'ltr' | 'rtl' };
+    getTextInfo: () => { direction: 'ltr' | 'rtl' };
     timeZones: string[];
     weekInfo: { firstDay: number };
     hourCycles: string[];
@@ -43,6 +43,9 @@ const USES_FAHRENHEIT = [
   'PW',
   'KY',
 ];
+
+// https://localizejs.com/articles/localizing-for-right-to-left-languages-the-issues-to-consider
+const USES_RTL = ['ar', 'arc', 'ckb', 'dv', 'fa', 'ha', 'he', 'khw', 'ks', 'ps', 'sd', 'ur', 'yi'];
 
 export function addLocaleListener(
   // NOTE(@kitten): We never use the event's data
@@ -72,16 +75,13 @@ export default {
   getLocales(): Locale[] {
     const locales = getNavigatorLocales();
     return locales?.map((languageTag) => {
-      // TextInfo is an experimental API that is not available in all browsers.
-      // We might want to consider using a locale lookup table instead.
-
       let locale = {} as ExtendedLocale;
 
       // Properties added only for compatibility with native, use `toLocaleString` instead.
       let digitGroupingSeparator: string | null = null;
       let decimalSeparator: string | null = null;
       let temperatureUnit: 'fahrenheit' | 'celsius' | null = null;
-
+      let textDirection: 'ltr' | 'rtl' | null = null;
       // Gracefully handle language codes like `en-GB-oed` which is unsupported
       // but is otherwise a valid language tag (grandfathered)
       try {
@@ -96,8 +96,9 @@ export default {
         }
       } catch {}
 
-      const { region, textInfo, language, script } = locale;
+      const { region, language, script } = locale;
 
+      textDirection = languageTextDirection(locale);
       if (region) {
         temperatureUnit = regionToTemperatureUnit(region);
       }
@@ -106,7 +107,7 @@ export default {
         languageTag,
         languageCode: language || languageTag.split('-')[0] || 'en',
         languageScriptCode: script || null,
-        textDirection: (textInfo?.direction as 'ltr' | 'rtl') || null,
+        textDirection,
         digitGroupingSeparator,
         decimalSeparator,
         measurementSystem: null,
@@ -138,4 +139,12 @@ export default {
 
 function regionToTemperatureUnit(region: string) {
   return USES_FAHRENHEIT.includes(region) ? 'fahrenheit' : 'celsius';
+}
+
+function languageTextDirection(locale: ExtendedLocale) {
+  // getTextInfo API is not available in all browsers.
+  if (typeof locale.getTextInfo === 'function') {
+    return locale.getTextInfo()?.direction;
+  }
+  return USES_RTL.includes(locale.language) ? 'rtl' : 'ltr';
 }